add TokenPriceService
